Crate dia

Library core.

Re-exports

pub use self::error::*;

pub use self::file::*;

pub use self::sci::*;

pub use self::sim::*;

pub use self::tools::*;

pub use self::util::*;

Modules

error
Error enumeration.

file
File module.

sci
Science module.

sim
Simulation module.

tools
Tools module.

util
Utilities module.

Macros

access
Succinctly create a reference method for the given variable.

args
Import command line arguments as a requested type.

clone
Succinctly create a clone method for the given variable.

map
Concisely construct a hashmap.

pause
Pause the thread for the given number of milliseconds.

report
Report an object and either its associated name, or a human readable string if supplied.

report_list
Report a list of items with an associated name, or a human readable string if supplied.