Trait dia::file::io::load::Load

pub trait Load where
    Self: Sized, {
    fn load(path: &Path) -> Result<Self, Error>;
}

Types implementing this trait can be loaded from a file.

Required methods

fn load(path: &Path) -> Result<Self, Error>

Deserialize the type from a given file.

Errors

if the target file can not be found, or the read string can not be serialised into an instance of the required type.

Loading content...